dia/diameter_3ggp_ts32_299.dia: Inherit RFC4005 instead of RFC7155
While it's true that RFC7155 obsolets RFC4005, latest release 18.0.1 of
3GPP TS 32.299 explicitly references RFC4005 and never RFC7155.
Since they are both incompatible, inherit the referenced RFC4005.

Same goes for 3GPP TS 29.273, but since we actually don't really use
it in the subset defined there, simply drop it.

As a result, nobody is using rfc7155.dia related stuff, so we can drop
it.

Change-Id: I7742d1523026e260b8501550977a41ebd3ad1b09